+++ examples/workflow/aggregates/user.commands.js
@@ -31,18 +31,15 @@ module.exports = {
     update: async (command, state/* , context */) => {
         if(!state.registered || state.removed)
             throw new ForbiddenError();
-        if(!command.payload.name && !command.payload.email)
-            throw new BadRequestError('Please specify a new name and/or email address');
-        if(command.payload.name === state.name)
-            throw new BadRequestError(`Name is already set to '${state.name}'`);
+        if(!command.payload.email)
+            throw new BadRequestError('Please specify a new email address');
         if(command.payload.email === state.email)
             throw new BadRequestError(`Email address is already set to '${state.email}'`);
 
         return {
             type: USER_UPDATED,
             payload: {
-                name: command.payload.name ?? state.name,
-                email: command.payload.email ?? state.email 
+                email: command.payload.email
             }
         };
     },