+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- MySQL read model schema types are now case insenitive
+
+### Fixed
+- Workflows no longer create a new state for every event in case of a replay
+- Bug where MySQL read model did not add indices after a schema change
+
+### Added
+- Support for variable-length strings in MySQL read model schema
 
 ## [1.1.0] - 2021-04-14
 ### Changed